+++ akshare/__init__.py
@@ -1936,9 +1936,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.5.14 add: add macro_china_bdti_index interface
 1.5.15 fix: fix bond_cov_jsl interface
 1.5.16 fix: fix stock_info_sh_delist interface
+1.5.17 add: add macro_china_bsi_index interface
 """
 
-__version__ = "1.5.16"
+__version__ = "1.5.17"
 __author__ = "AKFamily"
 
 import sys
@@ -3688,6 +3689,7 @@ from akshare.economic.macro_china import (
     macro_china_construction_price_index,
     macro_china_lpi_index,
     macro_china_bdti_index,
+    macro_china_bsi_index,
     macro_china_cpi_monthly,
     macro_china_cpi_yearly,
     macro_china_m2_yearly,