+++ docs/changelog.md
@@ -14,6 +14,10 @@
 
 ## 更新说明
 
+1.5.17 add: add macro_china_bsi_index interface
+
+    1. 修复 macro_china_bsi_index 接口，获取超灵便型船运价指数数据
+
 1.5.16 fix: fix stock_info_sh_delist interface
 
     1. 修复 stock_info_sh_delist 接口
@@ -320,6 +324,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.5.17 add: add macro_china_bsi_index interface
 
 1.5.16 fix: fix stock_info_sh_delist interface